requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
open_clip_torch
timm